requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.1
django-crispy-forms==1.9.0
django-debug-toolbar==2.1
django-taggit==1.2.0
django-tinymce4-lite==1.7.5
idna==2.9
jsmin==2.2.2
lxml==4.5.0
pathspec==0.6.0
Pillow==7.0.0
python-decouple==3.3
pytz==2019.3
regex==2019.12.19
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.8
